fw: align Probe RF to Edge-Node v2 (ADR-016) — RF RX test was dead
Three-codebase exploration of the provisioning toolchain found the Probe
(production test-jig firmware) still on RF v1 (sync 0x12, 13-byte frame)
while this session moved the Edge-Node to v2 (sync 0x4C, 14-byte, pack_id
uint16 CRC-16/CCITT). The Probe is the THIRD firmware that speaks the RF
protocol and was missed by the no-backward-compat lockstep — its frames
were rejected at the node PHY, so the provisioning station's RF-receiver
test (verifying a node's 433MHz receiver, its primary function) silently
produced no RX. CLAUDE.md's 2026-05-03 'RF RX count 持續增加' was true
then; v2 broke it.

Fix: RF_SYNC 0x12→0x4C, PROTO_FRAME_LEN 13→14, PROTO_VERSION=1 (flags[7:6]
=01), pack_id written as uint16 BE at bytes 11-12, CRC-8 moved to byte 13;
play()/_packId widened to uint16. Builds clean (pio run -e probe SUCCESS).

NOT flashed — needs the jig hardware + a node to confirm RF RX count
resumes. Companion change still needed in Provisioning-Station: its
pack_id is computed CRC-8 but v2 filters on CRC-16/CCITT, so the host must
pass the test pack's CRC-16 pack_id or the node filters the frames out.

+++ include/sx1262_tx.h
@@ -1,6 +1,6 @@
 #pragma once
 // ============================================================
-//  SX1262 TX — Send 13-byte playback frames to Node
+//  SX1262 TX — Send 14-byte v2 playback frames to Node (ADR-016)
 //  Same binary protocol as Hub (protocol.py)
 // ============================================================
 
@@ -16,7 +16,7 @@ public:
     void loop();   // Send periodic frame if playing
 
     // Commands (from UART SX: namespace)
-    void play(uint8_t seqIndex, uint8_t packId, uint8_t brightness = 100);
+    void play(uint8_t seqIndex, uint16_t packId, uint8_t brightness = 100);
     void stop();
     void seek(uint32_t progressMs);
     void setBrightness(uint8_t b);
@@ -34,7 +34,7 @@ private:
 
     bool     _playing = false;
     uint8_t  _seqIndex = 0;
-    uint8_t  _packId = 0;
+    uint16_t _packId = 0;
     uint8_t  _brightness = 100;
     uint32_t _progressMs = 0;
     uint32_t _playStartMs = 0;
